event_manager/requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==2.4.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==43.0.3
decorator==5.1.1
Django==5.1.2
django-celery-results==2.5.1
django-environ==0.11.2
djangorestframework==3.15.2
executing==2.1.0
ipdb==0.13.13
ipython==8.29.0
isort==5.13.2
jedi==0.19.1
kombu==5.4.2
matplotlib-inline==0.1.7
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.48
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
redis==5.2.0
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
